stabilize coverage decorations tests caused by race condition in event observe/trigger
