+++ src/providers/textDecorationProvider.ts
@@ -1,7 +1,7 @@
 import { DecorationOptions, Event, ProviderResult, Range, TextEditor, TextEditorDecorationType } from "vscode";
 
 export interface TextDecorationProvider {
-    onDecorationsChanged?: Event<TextEditor[] | undefined>;
+    onDecorationsChanged?: Event<readonly TextEditor[] | undefined>;
     decorationType: TextEditorDecorationType;
     getDecorations(editor: TextEditor): ProviderResult<Range[] | DecorationOptions[]>;
 }
