+++ src/test/types.ts
@@ -1,4 +1,5 @@
 export const TEST_TYPES = {
     GitExtension: Symbol.for("GitExtensionStub"),
-    Logger: Symbol.for("Logger")
+    Logger: Symbol.for("Logger"),
+    CoverageDecorationsObserver: Symbol.for("CoverageDecorationsObserver"),
 };
